# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /storage/uploads/finalize:
    post:
      tags: [storage]
      operationId: finalizeUpload
      summary: Mark an uploaded object as public so it is servable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeUploadInput'
      responses:
        '200':
          description: Upload finalized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeUploadOutput'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500':
          description: Failed to finalize upload.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema: { type: string, format: binary }
        '404':
          description: Not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ── Public catalog ────────────────────────────────────────────────────
...
  /admin/products:
    get:
      tags: [admin]
      operationId: adminListProducts
      summary: List all products (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    post:
      tags: [admin]
      operationId: adminCreateProduct
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /admin/products/reorder:
    post:
      tags: [admin]
      operationId: adminReorderProducts
      summary: Reorder products by id list (index becomes sortOrder)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReorderRequest" }
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /admin/products/{id}:
    patch:
      tags: [admin]
      operationId: adminUpdateProduct
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    delete:
      tags: [admin]
      operationId: adminDeleteProduct
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ── Admin: Articles ───────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata: { $ref: '#/components/schemas/UploadUrlRequest' }
    FinalizeUploadInput:
      type: object
      required: [objectPath]
      properties:
        objectPath: { type: string, minLength: 1 }
    FinalizeUploadOutput:
      type: object
      required: [objectPath]
      properties:
        objectPath: { type: string }
    Variant:
      type: object
      required: [size, price, stock]
      properties:
        size: { type: string }
        price: { type: integer }
        stock: { type: integer }
    BrewingGuide:
      type: object
      required: [temperature, steepTime, cupSize, teaAmount, steps]
      properties:
        temperature: { type: string }
        steepTime: { type: string }
        cupSize: { type: string }
        teaAmount: { type: string }
        steps:
          type: array
          items: { type: string }
    Product:
      type: object
      required:
        - id
        - slug
        - name
        - category
        - description
        - shortDescription
        - price
        - rating
        - reviewCount
        - variants
        - tastingNotes
        - ingredients
        - wellnessBenefits
        - brewingGuide
        - origin
        - harvestSeason
        - caffeineLevel
        - flavorProfile
        - wellnessFocus
        - brewingType
        - ritualStyle
        - imageUrl
        - categoryTheme
        - pairsWith
        - moodTags
        - plantationStory
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        shortDescription: { type: string }
        price: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        tastingNotes:
          type: array
          items: { type: string }
        ingredients:
          type: array
          items: { type: string }
        wellnessBenefits:
          type: array
          items: { type: string }
        brewingGuide: { $ref: "#/components/schemas/BrewingGuide" }
        origin: { type: string }
        harvestSeason: { type: string }
        caffeineLevel:
          type: string
          enum: [none, low, medium, high]
        flavorProfile:
          type: array
          items: { type: string }
        wellnessFocus:
          type: array
          items: { type: string }
        brewingType: { type: string }
        ritualStyle: { type: string }
        fomoTag:
          type: string
          nullable: true
        imageUrl: { type: string }
        images:
          type: array
          items: { $ref: "#/components/schemas/ProductImage" }
        categoryTheme: { $ref: "#/components/schemas/CategoryTheme" }
        pairsWith:
          type: array
          items: { type: string }
        moodTags:
          type: array
          items: { type: string }
        plantationStory: { type: string }
        metaTitle: { type: string }
        metaDescription: { type: string }
        jsonLd:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: "null" }
        relatedProductIds:
          type: array
          items: { type: string }
        relatedArticleSlugs:
          type: array
          items: { type: string }
    ProductImage:
      type: object
      required: [url, kind]
      properties:
        url: { type: string }
        kind:
          type: string
          enum: [hero, studio]
        alt: { type: string }
        isAi: { type: boolean }
...
    CategoryTheme:
      type: object
      required: [primary, accent, bg, gradient, dark]
      properties:
        primary: { type: string }
        accent: { type: string }
        bg: { type: string }
        gradient: { type: string }
        dark: { type: string }
    ProductInput:
      allOf:
        - $ref: "#/components/schemas/Product"
...
    AdminReorderRequest:
      type: object
      required: [order]
      properties:
        order:
          type: array
          items: { type: string }
...